fix: :adhesive_bandage: Improve types used and make use of `saturating` methods

+++ backend/lib/src/models/files.rs
@@ -50,9 +50,9 @@ pub struct FileInfo {
     #[serde(rename = "txHash", skip_serializing_if = "Option::is_none")]
     pub tx_hash: Option<String>,
     #[serde(rename = "desiredReplicas")]
-    pub desired_replicas: i32,
+    pub desired_replicas: u32,
     #[serde(rename = "currentReplication")]
-    pub current_replication: i64,
+    pub current_replication: u32,
 }
 
 impl FileInfo {
@@ -75,8 +75,8 @@ impl FileInfo {
     pub fn from_db(
         db: &DBFile,
         is_public: bool,
-        desired_replicas: i32,
-        current_replication: i64,
+        desired_replicas: u32,
+        current_replication: u32,
     ) -> Self {
         Self {
             file_key: hex::encode(&db.file_key),
